pyproject.toml
[tool.poetry]
name = "hyperparameter_optimization"
version = "0.1.0"
description = ""
authors = ["FugaTakata <fuga.takata.dev@gmail.com>"]
readme = "README.md"
packages = [
  { include = "config", from = "src" },
  { include = "data_processors", from = "src" },
  { include = "generators", from = "src" },
  { include = "layouts", from = "src" },
  { include = "optimizers", from = "src" },
  { include = "quality_metrics", from = "src" },
  { include = "utils", from = "src" },
]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
networkx = { extras = ["default"], version = "^3.0" }
python-dotenv = "^0.21.1"
tqdm = "^4.64.1"
pandas = "^1.5.3"
matplotlib = "^3.7.0"
optuna = "^3.1.0"
numpy = "^1.24.2"
scipy = "^1.10.1"
scikit-learn = "^1.2.1"
optuna-dashboard = "^0.8.1"
plotly = "^5.13.0"
nbformat = "^5.7.3"
kaleido = "0.2.1"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pyproject-flake8 = "^6.0.0.post1"
maturin = "^0.14.13"
ipykernel = "^6.21.2"


[tool.poetry.group.notification.dependencies]
discord-py = "^2.1.1"
aiohttp = "^3.8.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79
import_heading_stdlib = "Standard Library"
import_heading_thirdparty = "Third Party Library"
import_heading_firstparty = "First Party Library"
import_heading_localfolder = "Local Library"
multi_line_output = 3
include_trailing_comma = true
known_firstparty = [
  "config",
  "data_processors",
  "generators",
  "layouts",
  "optimizers",
  "quality_metrics",
  "utils",
]

[tool.flake8]
max-line-length = 79
exclude = [".venv", ".git", "__pycache__"]
ignore = ['E203', 'E501', 'W503']
max-complexity = 10
